refactor(core): Move atomic style generation entirely in-house.
In preparing for big improvements to bundle size and runtime
performance, moving the atomic CSS generator into the repo.

This change also begins optimization by reducing one of the loops.

It unlocks removing quite a few more loops next, as well as being able
to fully separate the react-native dependency in future versions.